+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Runt Changelog
 0.1.2
 -----
 - Execute all commands in the directory where `runt.toml` resides.
+- Print out test suite name and the total number of test states.
+- Return code is the number of failing tests.
 
 0.1.1
 -----